refactor(layout): unify point_to_bin_index output dtype on np.intp
The six `point_to_bin_index` implementations returned divergent index dtypes
(`int32` in the KD-tree/grid mixins, `int64` in the transformed-layout shim,
`int` in hexagonal) behind a `LayoutEngine` contract annotated `np.int_`.
Normalize all of them — and the base contract — on `np.intp` (pointer-sized:
64-bit on every platform, including Windows). The internal
`_kdtree_nodes_to_bin_indices_map` lookup keeps its `int32` storage; its values
widen into the `intp` output on assignment.

No test pinned the layout-level dtype, and output values are unchanged (a lossless
int widening). mypy neutral on default and `--platform win32`; 5105 tests pass.

+++ src/neurospatial/layout/engines/triangular_mesh.py
@@ -183,7 +183,7 @@ class TriangularMeshLayout:
         self.active_mask[active_original_indices] = True
         # self.grid_edges remains an empty tuple as it's not a rectilinear grid.
 
-    def point_to_bin_index(self, points: NDArray[np.float64]) -> NDArray[np.int_]:
+    def point_to_bin_index(self, points: NDArray[np.float64]) -> NDArray[np.intp]:
         """Map each 2D point to an active triangle index, or -1 if outside.
 
         Uses Delaunay.find_simplex() → original-simplex index → active-triangle index via
@@ -197,7 +197,7 @@ class TriangularMeshLayout:
 
         Returns
         -------
-        NDArray[np.int_]
+        NDArray[np.intp]
             Each entry is in [0..n_active-1] or -1 if outside.
 
         """
@@ -243,7 +243,7 @@ class TriangularMeshLayout:
                     if not keep:
                         active_triangle_idxs[idxs[local_i]] = -1
 
-        return active_triangle_idxs
+        return active_triangle_idxs.astype(np.intp, copy=False)
 
     @property
     def is_linearized_track(self) -> bool:
